almost there continue for teams switching broken
