+++ core/protocol/passThrough.ts
@@ -82,7 +82,5 @@ export const CORE_TO_WEBVIEW_PASS_THROUGH: (keyof ToWebviewFromCoreProtocol)[] =
     "setTTSActive",
     "getWebviewHistoryLength",
     "getCurrentSessionId",
-    "signInToControlPlane",
-    "openDialogMessage",
     "docs/suggestions",
   ];
