+++ extensions/vscode/src/commands.ts
@@ -916,7 +916,8 @@ const getCommandsMap: (
         },
         {
           label: "$(screen-full) Open full screen chat",
-          description: getMetaKeyLabel() + " + K, " + getMetaKeyLabel() + " + M",
+          description:
+            getMetaKeyLabel() + " + K, " + getMetaKeyLabel() + " + M",
         },
         {
           label: quickPickStatusText(targetStatus),
@@ -997,12 +998,6 @@ const getCommandsMap: (
     "continue.navigateTo": (path: string, toggle: boolean) => {
       sidebar.webviewProtocol?.request("navigateTo", { path, toggle });
       focusGUI();
-    },
-    "continue.signInToControlPlane": () => {
-      sidebar.webviewProtocol?.request("signInToControlPlane", undefined);
-    },
-    "continue.openAccountDialog": () => {
-      sidebar.webviewProtocol?.request("openDialogMessage", "account");
     },
     "continue.startLocalOllama": () => {
       startLocalOllama(ide);
