+++ gui/src/components/modelSelection/platform/shared.tsx
@@ -8,6 +8,7 @@ import { useState } from "react";
 import styled from "styled-components";
 import { defaultBorderRadius, lightGray } from "../..";
 import { ToolTip } from "../../gui/Tooltip";
+import { ProfileDescription } from "core/config/ConfigHandler";
 
 export const OptionDiv = styled.div<{
   isDisabled?: boolean;
@@ -54,8 +55,7 @@ interface ModelOptionProps {
   disabled: boolean;
   selected: boolean;
   showConfigure: boolean;
-  onLink?: (e: any) => void;
-  onConfigure: (e: any) => void;
+  onOpenConfig: () => void;
   onClick: () => void;
   errors?: ConfigValidationError[];
   onClickError?: (e: any) => void;
@@ -85,16 +85,15 @@ const StyledExclamationTriangleIcon = styled(IconBase).attrs({
 })``;
 
 export function Option({
-  onConfigure,
   children,
   idx,
   disabled,
   onClick,
   showConfigure,
-  onLink,
   selected,
   errors,
   onClickError,
+  onOpenConfig,
 }: ModelOptionProps) {
   const [hovered, setHovered] = useState(false);
 
@@ -119,16 +118,27 @@ export function Option({
         <div className="flex w-full items-center justify-between">
           {children}
           <div className="ml-2 flex items-center">
-            {!errors?.length && showConfigure && (
-              <StyledCog6ToothIcon $hovered={hovered} onClick={onConfigure} />
-            )}
-            {!errors?.length && onLink && (
-              <StyledArrowTopRightOnSquareIcon
-                $hovered={hovered}
-                onClick={onLink}
-              />
-            )}
-            {errors?.length ? (
+            {!errors?.length ? (
+              showConfigure ? (
+                <StyledCog6ToothIcon
+                  $hovered={hovered}
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    e.preventDefault();
+                    onOpenConfig();
+                  }}
+                />
+              ) : (
+                <StyledArrowTopRightOnSquareIcon
+                  $hovered={hovered}
+                  onClick={(e) => {
+                    e.stopPropagation();
+                    e.preventDefault();
+                    onOpenConfig();
+                  }}
+                />
+              )
+            ) : (
               <>
                 <StyledExclamationTriangleIcon
                   data-tooltip-id={`${idx}-errors-tooltip`}
@@ -141,7 +151,7 @@ export function Option({
                   {JSON.stringify(errors, null, 2)}
                 </ToolTip>
               </>
-            ) : null}
+            )}
           </div>
         </div>
       </div>
