+++ gui/src/pages/config/index.tsx
@@ -1,23 +1,17 @@
-import { useContext, useEffect, useState } from "react";
+import { Fragment, useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import PageHeader from "../../components/PageHeader";
 import { IdeMessengerContext } from "../../context/IdeMessenger";
 import { useNavigationListener } from "../../hooks/useNavigationListener";
-import { useAppDispatch, useAppSelector } from "../../redux/hooks";
-
-import { CheckIcon, XMarkIcon } from "@heroicons/react/24/outline";
-import {
-  SharedConfigSchema,
-  modifyContinueConfigWithSharedConfig,
-} from "core/config/sharedConfig";
-import { Input, SecondaryButton } from "../../components";
-import NumberInput from "../../components/gui/NumberInput";
-import { Select } from "../../components/gui/Select";
-import ToggleSwitch from "../../components/gui/Switch";
+import { useAppDispatch } from "../../redux/hooks";
+import { Button, SecondaryButton } from "../../components";
 import { useAuth } from "../../context/Auth";
-import { updateConfig } from "../../redux/slices/configSlice";
-import { getFontSize } from "../../util";
 import { ScopeSelect } from "./ScopeSelect";
+import UserSettingsUI from "./UserSettings";
+import { Listbox, Transition } from "@headlessui/react";
+import { ChevronUpDownIcon } from "@heroicons/react/24/outline";
+import { setSelectedProfileId } from "../../redux/slices/sessionSlice";
+import { isNull } from "lodash";
 
 function ConfigPage() {
   useNavigationListener();
@@ -25,81 +19,29 @@ function ConfigPage() {
   const navigate = useNavigate();
   const ideMessenger = useContext(IdeMessengerContext);
 
-  const { selectedProfile, selectedOrganization, session } = useAuth();
-  const config = useAppSelector((state) => state.config.config);
+  const {
+    session,
 
-  function handleUpdate(sharedConfig: Partial<SharedConfigSchema>) {
-    dispatch(
-      updateConfig(modifyContinueConfigWithSharedConfig(config, sharedConfig)),
-    );
-    ideMessenger.post("config/updateSharedConfig", sharedConfig);
-  }
+    logout,
+    login,
+    profiles,
+    selectedProfile,
+    controlServerBetaEnabled,
+    selectedOrganization,
+  } = useAuth();
+  // const profiles = [];
 
-  const codeWrap = config.ui?.codeWrap ?? false;
-  const showChatScrollbar = config.ui?.showChatScrollbar ?? false;
-  const displayRawMarkdown = config.ui?.displayRawMarkdown ?? false;
-  const disableSessionTitles = config.disableSessionTitles ?? false;
-  const readResponseTTS = config.experimental?.readResponseTTS ?? false;
-
-  const allowAnonymousTelemetry = config.allowAnonymousTelemetry ?? true;
-  const disableIndexing = config.disableIndexing ?? false;
-
-  const useAutocompleteCache = config.tabAutocompleteOptions?.useCache ?? false;
-  const useChromiumForDocsCrawling =
-    config.experimental?.useChromiumForDocsCrawling ?? false;
-  const codeBlockToolbarPosition = config.ui?.codeBlockToolbarPosition ?? "top";
-  const useAutocompleteMultilineCompletions =
-    config.tabAutocompleteOptions?.multilineCompletions ?? "auto";
-  const fontSize = getFontSize();
+  const changeProfileId = (id: string) => {
+    ideMessenger.post("didChangeSelectedProfile", { id });
+    dispatch(setSelectedProfileId(id));
+  };
 
   const [hubEnabled, setHubEnabled] = useState(false);
-
   useEffect(() => {
     ideMessenger.ide.getIdeSettings().then(({ continueTestEnvironment }) => {
       setHubEnabled(continueTestEnvironment === "production");
     });
   }, [ideMessenger]);
-
-  // Disable autocomplete
-  const disableAutocompleteInFiles = (
-    config.tabAutocompleteOptions?.disableInFiles ?? []
-  ).join(", ");
-  const [formDisableAutocomplete, setFormDisableAutocomplete] = useState(
-    disableAutocompleteInFiles,
-  );
-  const cancelChangeDisableAutocomplete = () => {
-    setFormDisableAutocomplete(disableAutocompleteInFiles);
-  };
-  const handleDisableAutocompleteSubmit = () => {
-    handleUpdate({
-      disableAutocompleteInFiles: formDisableAutocomplete
-        .split(",")
-        .map((val) => val.trim())
-        .filter((val) => !!val),
-    });
-  };
-
-  useEffect(() => {
-    // Necessary so that reformatted/trimmed values don't cause dirty state
-    setFormDisableAutocomplete(disableAutocompleteInFiles);
-  }, [disableAutocompleteInFiles]);
-
-  // Workspace prompts
-  const promptPath = config.experimental?.promptPath || "";
-  const [formPromptPath, setFormPromptPath] = useState(promptPath);
-  const cancelChangePromptPath = () => {
-    setFormPromptPath(promptPath);
-  };
-  const handleSubmitPromptPath = () => {
-    handleUpdate({
-      promptPath: formPromptPath || "",
-    });
-  };
-
-  useEffect(() => {
-    // Necessary so that reformatted/trimmed values don't cause dirty state
-    setFormPromptPath(promptPath);
-  }, [promptPath]);
 
   function handleOpenConfig() {
     if (!selectedProfile) {
@@ -110,301 +52,156 @@ function ConfigPage() {
     });
   }
 
-  if (!selectedProfile) {
-    return (
-      <div className="overflow-y-scroll">
-        <PageHeader onTitleClick={() => navigate("/")} title="Chat" />
-        <div className="px-4">
-          <div>
-            <h2>Continue Config</h2>
-          </div>
-          <p className="italic">No config profile selected</p>
-        </div>
-      </div>
-    );
-  }
+  // NOTE Hub takes priority over Continue for Teams
+  // Since teams will be moving to hub, not vice versa
 
   return (
     <div className="overflow-y-scroll">
       <PageHeader onTitleClick={() => navigate("/")} title="Chat" />
 
       <div className="divide-x-0 divide-y-2 divide-solid divide-zinc-700 px-4">
-        {hubEnabled && !!session && (
-          <div className="flex flex-col gap-4 py-6">
-            <div>
-              <h2 className="mb-1 mt-0">Account</h2>
-              {selectedOrganization?.name && (
-                <span className="text-lightgray">
-                  You are currently signed in to{" "}
-                  <span className="font-semibold">
-                    {selectedOrganization?.name}
+        <div className="flex flex-col">
+          <div className="flex max-w-[400px] flex-col gap-4 py-4">
+            <h2 className="mb-1 mt-0">Account</h2>
+            {!session ? (
+              <div className="flex flex-col gap-2">
+                <span>You are not signed in.</span>
+                <SecondaryButton onClick={() => login(false)}>
+                  Sign in
+                </SecondaryButton>
+              </div>
+            ) : (
+              <div className="flex flex-col gap-4">
+                {hubEnabled ? (
+                  // Hub: show org selector
+                  <div className="flex flex-col gap-1.5">
+                    <span className="text-lightgray">{`Organization`}</span>
+                    <ScopeSelect />
+                  </div>
+                ) : (
+                  // Continue for teams: show org text
+                  <div>You are using Continue for Teams</div>
+                )}
+                <div className="flex flex-row items-center gap-2">
+                  <span className="text-lightgray">
+                    {session.account.label === ""
+                      ? "Signed in"
+                      : `Signed in as ${session.account.label}`}
                   </span>
-                </span>
+                  <span
+                    onClick={logout}
+                    className="text-lightgray cursor-pointer underline"
+                  >{`Sign out`}</span>
+                </div>
+              </div>
+            )}
+          </div>
+        </div>
+
+        <div className="flex flex-col">
+          <div className="flex max-w-[400px] flex-col gap-4 py-6">
+            <h2 className="mb-1 mt-0">Configuration</h2>
+            <Listbox value={selectedProfile?.id} onChange={changeProfileId}>
+              {({ open }) => (
+                <div className="relative w-full">
+                  <Listbox.Button className="border-vsc-input-border bg-vsc-background hover:bg-vsc-input-background text-vsc-foreground relative m-0 flex w-full cursor-pointer items-center justify-between rounded-md border border-solid px-3 py-2 text-left">
+                    <span className="lines lines-1">
+                      {selectedProfile?.title}
+                    </span>
+                    <div className="pointer-events-none flex items-center">
+                      <ChevronUpDownIcon
+                        className="h-5 w-5"
+                        aria-hidden="true"
+                      />
+                    </div>
+                  </Listbox.Button>
+
+                  <Transition
+                    as={Fragment}
+                    show={open}
+                    enter="transition ease-out duration-100"
+                    enterFrom="transform opacity-0 scale-95"
+                    enterTo="transform opacity-100 scale-100"
+                    leave="transition ease-in duration-75"
+                    leaveFrom="transform opacity-100 scale-100"
+                    leaveTo="transform opacity-0 scale-95"
+                  >
+                    <Listbox.Options className="bg-vsc-background max-h-80vh absolute mt-0.5 w-full overflow-y-scroll rounded-sm p-0">
+                      {profiles.map((option, idx) => (
+                        <Listbox.Option
+                          key={idx}
+                          value={option.id}
+                          className={`text-vsc-foreground hover:text-vsc-list-active-foreground flex cursor-pointer flex-row items-center gap-3 px-3 py-2 ${selectedProfile?.id === option.id ? "bg-vsc-list-active-background" : "bg-vsc-input-background"}`}
+                        >
+                          <span className="lines lines-1 relative flex h-5 items-center justify-between gap-3 pr-2 text-xs">
+                            {option.title}
+                          </span>
+                        </Listbox.Option>
+                      ))}
+                      {profiles.length === 0 && (
+                        <Listbox.Option
+                          key={"no-profiles"}
+                          value={undefined}
+                          className={`text-vsc-foreground hover:text-vsc-list-active-foreground flex cursor-pointer flex-row items-center gap-3 px-3 py-2 ${selectedProfile?.id === undefined ? "bg-vsc-list-active-background" : "bg-vsc-input-background"}`}
+                        >
+                          <span className="lines lines-1 relative flex h-5 items-center justify-between gap-3 pr-2 text-xs italic">
+                            No profiles found
+                          </span>
+                        </Listbox.Option>
+                      )}
+                    </Listbox.Options>
+                  </Transition>
+                </div>
               )}
+            </Listbox>
+            {selectedProfile && (
+              <SecondaryButton onClick={handleOpenConfig}>
+                {selectedProfile.id === "local"
+                  ? "Open Config File"
+                  : hubEnabled
+                    ? "Open Assistant"
+                    : "Open Workspace"}
+              </SecondaryButton>
+            )}
+          </div>
+        </div>
+        {!controlServerBetaEnabled || hubEnabled ? (
+          <div className="flex flex-col">
+            <div className="flex max-w-[400px] flex-col">
+              <UserSettingsUI />
             </div>
-
-            <ScopeSelect />
           </div>
-        )}
-
-        <div className="flex flex-col gap-4 py-6">
-          <h2 className="mb-1 mt-0">Configuration</h2>
-
-          <SecondaryButton
-            onClick={handleOpenConfig}
-            className="!my-0 max-w-[400px]"
-          >
-            Open config file
-          </SecondaryButton>
-        </div>
-
-        <div className="flex flex-col gap-4 py-6">
-          <div>
-            <h2 className="mb-2 mt-0">User settings</h2>
-          </div>
-
-          <div className="flex flex-col gap-4">
-            <ToggleSwitch
-              isToggled={codeWrap}
-              onToggle={() =>
-                handleUpdate({
-                  codeWrap: !codeWrap,
-                })
-              }
-              text="Wrap Codeblocks"
-            />
-            <ToggleSwitch
-              isToggled={displayRawMarkdown}
-              onToggle={() =>
-                handleUpdate({
-                  displayRawMarkdown: !displayRawMarkdown,
-                })
-              }
-              text="Display Raw Markdown"
-            />
-            <ToggleSwitch
-              isToggled={allowAnonymousTelemetry}
-              onToggle={() =>
-                handleUpdate({
-                  allowAnonymousTelemetry: !allowAnonymousTelemetry,
-                })
-              }
-              text="Allow Anonymous Telemetry"
-            />
-            <ToggleSwitch
-              isToggled={disableIndexing}
-              onToggle={() =>
-                handleUpdate({
-                  disableIndexing: !disableIndexing,
-                })
-              }
-              text="Disable Indexing"
-            />
-
-            <ToggleSwitch
-              isToggled={disableSessionTitles}
-              onToggle={() =>
-                handleUpdate({
-                  disableSessionTitles: !disableSessionTitles,
-                })
-              }
-              text="Disable Session Titles"
-            />
-            <ToggleSwitch
-              isToggled={readResponseTTS}
-              onToggle={() =>
-                handleUpdate({
-                  readResponseTTS: !readResponseTTS,
-                })
-              }
-              text="Response Text to Speech"
-            />
-
-            <ToggleSwitch
-              isToggled={showChatScrollbar}
-              onToggle={() =>
-                handleUpdate({
-                  showChatScrollbar: !showChatScrollbar,
-                })
-              }
-              text="Show Chat Scrollbar"
-            />
-
-            <ToggleSwitch
-              isToggled={useAutocompleteCache}
-              onToggle={() =>
-                handleUpdate({
-                  useAutocompleteCache: !useAutocompleteCache,
-                })
-              }
-              text="Use Autocomplete Cache"
-            />
-
-            <ToggleSwitch
-              isToggled={useChromiumForDocsCrawling}
-              onToggle={() =>
-                handleUpdate({
-                  useChromiumForDocsCrawling: !useChromiumForDocsCrawling,
-                })
-              }
-              text="Use Chromium for Docs Crawling"
-            />
-
-            <label className="flex items-center justify-between gap-3">
-              <span className="text-right">Codeblock Actions Position</span>
-              <Select
-                value={codeBlockToolbarPosition}
-                onChange={(e) =>
-                  handleUpdate({
-                    codeBlockToolbarPosition: e.target.value as
-                      | "top"
-                      | "bottom",
-                  })
-                }
-              >
-                <option value="top">Top</option>
-                <option value="bottom">Bottom</option>
-              </Select>
-            </label>
-
-            <label className="flex items-center justify-between gap-3">
-              <span className="text-right">Multiline Autocompletions</span>
-              <Select
-                value={useAutocompleteMultilineCompletions}
-                onChange={(e) =>
-                  handleUpdate({
-                    useAutocompleteMultilineCompletions: e.target.value as
-                      | "auto"
-                      | "always"
-                      | "never",
-                  })
-                }
-              >
-                <option value="auto">Auto</option>
-                <option value="always">Always</option>
-                <option value="never">Never</option>
-              </Select>
-            </label>
-
-            <label className="flex items-center justify-between gap-3">
-              <span className="text-right">Font Size</span>
-              <NumberInput
-                value={fontSize}
-                onChange={(val) =>
-                  handleUpdate({
-                    fontSize: val,
-                  })
-                }
-                min={7}
-                max={50}
-              />
-            </label>
-
-            <label className="flex items-center justify-between gap-3">
-              <span className="text-right">Font Size</span>
-              <NumberInput
-                value={fontSize}
-                onChange={(val) =>
-                  handleUpdate({
-                    fontSize: val,
-                  })
-                }
-                min={7}
-                max={50}
-              />
-            </label>
-            <form
-              className="flex flex-col gap-1"
-              onSubmit={(e) => {
-                e.preventDefault();
-                handleSubmitPromptPath();
-              }}
-            >
-              <div className="flex items-center justify-between">
-                <span>Workspace prompts path</span>
-                <div className="flex items-center gap-2">
-                  <Input
-                    value={formPromptPath}
-                    className="max-w-[100px]"
-                    onChange={(e) => {
-                      setFormPromptPath(e.target.value);
-                    }}
-                  />
-                  <div className="flex h-full flex-col">
-                    {formPromptPath !== promptPath ? (
-                      <>
-                        <div
-                          onClick={handleSubmitPromptPath}
-                          className="cursor-pointer"
-                        >
-                          <CheckIcon className="h-4 w-4 text-green-500 hover:opacity-80" />
-                        </div>
-                        <div
-                          onClick={cancelChangePromptPath}
-                          className="cursor-pointer"
-                        >
-                          <XMarkIcon className="h-4 w-4 text-red-500 hover:opacity-80" />
-                        </div>
-                      </>
-                    ) : (
-                      <div>
-                        <CheckIcon className="text-vsc-foreground-muted h-4 w-4" />
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-            </form>
-            <form
-              className="flex flex-col gap-1"
-              onSubmit={(e) => {
-                e.preventDefault();
-                handleDisableAutocompleteSubmit();
-              }}
-            >
-              <div className="flex items-center justify-between">
-                <span>Disable autocomplete in files</span>
-                <div className="flex items-center gap-2">
-                  <Input
-                    value={formDisableAutocomplete}
-                    className="max-w-[100px]"
-                    onChange={(e) => {
-                      setFormDisableAutocomplete(e.target.value);
-                    }}
-                  />
-                  <div className="flex h-full flex-col">
-                    {formDisableAutocomplete !== disableAutocompleteInFiles ? (
-                      <>
-                        <div
-                          onClick={handleDisableAutocompleteSubmit}
-                          className="cursor-pointer"
-                        >
-                          <CheckIcon className="h-4 w-4 text-green-500 hover:opacity-80" />
-                        </div>
-                        <div
-                          onClick={cancelChangeDisableAutocomplete}
-                          className="cursor-pointer"
-                        >
-                          <XMarkIcon className="h-4 w-4 text-red-500 hover:opacity-80" />
-                        </div>
-                      </>
-                    ) : (
-                      <div>
-                        <CheckIcon className="text-vsc-foreground-muted h-4 w-4" />
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
-              <span className="text-vsc-foreground-muted self-end text-xs">
-                Comma-separated list of path matchers
-              </span>
-            </form>
-          </div>
-        </div>
+        ) : null}
       </div>
     </div>
   );
 }
 
 export default ConfigPage;
+
+// export default function AccountDialog() {
+//   const ideMessenger = useContext(IdeMessengerContext);
+
+//   const dispatch = useDispatch();
+
+//   return (
+//     <div className="flex flex-col items-center gap-5 p-4 pt-0">
+//       <div className="flex flex-col items-center text-center">
+//         <h1 className="mb-1 text-2xl">Account</h1>
+//         <span className="text-lightgray text-sm">
+
+//         </span>
+//       </div>
+
+//       {controlServerBetaEnabled && (
+//         <div className="mb-10 mt-4 w-full">
+//           <span className="mb-1 block text-sm">Current workspace</span>
+
+//         </div>
+//       )}
+
+//
+//       </div>
+//     </div>
+//   );
+// }
